Extracting ServiceControlApis onto their own project
