+++ src/ServiceControl.Api/IConfigurationApi.cs
@@ -1,0 +1,13 @@
+﻿namespace ServiceControl.Api
+{
+    using System.Threading;
+    using System.Threading.Tasks;
+    using Contracts;
+
+    public interface IConfigurationApi
+    {
+        public Task<RootUrls> GetUrls(string baseUrl, CancellationToken cancellationToken);
+        public Task<object> GetConfig(CancellationToken cancellationToken);
+        public Task<RemoteConfiguration[]> GetRemoteConfigs(CancellationToken cancellationToken);
+    }
+}
