+++ src/ServiceControl.Api/IEndpointsApi.cs
@@ -1,0 +1,12 @@
+﻿namespace ServiceControl.Api
+{
+    using System.Collections.Generic;
+    using System.Threading;
+    using System.Threading.Tasks;
+    using ServiceControl.Api.Contracts;
+
+    public interface IEndpointsApi
+    {
+        public Task<List<Endpoint>> GetEndpoints(CancellationToken cancellationToken);
+    }
+}
