+++ src/ServiceControl.UnitTests/API/APIApprovals.cs
@@ -15,6 +15,7 @@
     using Particular.Approvals;
     using Particular.ServiceControl.Licensing;
     using ServiceBus.Management.Infrastructure.Settings;
+    using ServiceControl.Infrastructure.Api;
     using ServiceControl.Infrastructure.WebApi;
 
     [TestFixture]
@@ -27,11 +28,12 @@
             var actionContext = new ActionContext { HttpContext = httpContext, RouteData = new RouteData(), ActionDescriptor = new ControllerActionDescriptor() };
             var controllerContext = new ControllerContext(actionContext);
 
-            var controller = new RootController(
-                new ActiveLicense { IsValid = true },
-                new Settings(),
-                httpClientFactory: null
+            var controller = new RootController(new ConfigurationApi(
+                    new ActiveLicense { IsValid = true },
+                    new Settings(),
+                    null
                 )
+            )
             {
                 ControllerContext = controllerContext,
                 Url = new UrlHelper(actionContext)
