+++ src/ServiceControl.UnitTests/ApprovalFiles/APIApprovals.HttpApiRoutes.approved.txt
@@ -1,7 +1,7 @@
 GET / => ServiceControl.Infrastructure.WebApi.RootController:Urls()
 GET /archive/groups/id/{groupId:required:minlength(1)} => ServiceControl.MessageFailures.Api.ArchiveMessagesController:GetGroup(String groupId, String status, String modified)
 GET /configuration => ServiceControl.Infrastructure.WebApi.RootController:Config()
-GET /configuration/remotes => ServiceControl.Infrastructure.WebApi.RootController:RemoteConfig()
+GET /configuration/remotes => ServiceControl.Infrastructure.WebApi.RootController:RemoteConfig(CancellationToken cancellationToken)
 GET /connection => ServiceControl.Connection.ConnectionController:GetConnectionDetails()
 GET /conversations/{conversationId:required:minlength(1)} => ServiceControl.CompositeViews.Messages.GetMessagesByConversationController:Messages(PagingInfo pagingInfo, SortInfo sortInfo, Boolean includeSystemMessages, String conversationId)
 GET /customchecks => ServiceControl.CustomChecks.CustomCheckController:CustomChecks(PagingInfo pagingInfo, String status)
