+++ src/ServiceControl/ServiceControl.csproj
@@ -16,6 +16,7 @@
   <Import Project="..\ProjectReferences.Transports.props" />
 
   <ItemGroup>
+    <ProjectReference Include="..\ServiceControl.Api\ServiceControl.Api.csproj"/>
     <ProjectReference Include="..\ServiceControl.Configuration\ServiceControl.Configuration.csproj" />
     <ProjectReference Include="..\ServiceControl.Infrastructure.Metrics\ServiceControl.Infrastructure.Metrics.csproj" />
     <ProjectReference Include="..\ServiceControl.Infrastructure\ServiceControl.Infrastructure.csproj" />
